fix(mobile): scope selectable text to Android
