+++ apps/mobile/src/native/SelectableMarkdownText.tsx
@@ -1,9 +1,4 @@
-import {
-  SelectableMarkdownText as T3SelectableMarkdownText,
-  type SelectableMarkdownTextProps,
-} from "@t3tools/mobile-markdown-text/renderer";
-
-import { highlightCodeSnippet } from "../features/review/shikiReviewHighlighter";
+import type { SelectableMarkdownTextProps } from "@t3tools/mobile-markdown-text/renderer";
 
 type MobileSelectableMarkdownTextProps = Omit<SelectableMarkdownTextProps, "highlightCode">;
 
@@ -16,12 +11,10 @@ export type {
   SelectableMarkdownSkill,
 } from "@t3tools/mobile-markdown-text/types";
 
-// The renderer falls back to React Native Text outside iOS, so Android can use
-// the same Markdown chunking while retaining native text selection.
 export function hasNativeSelectableMarkdownText(): boolean {
-  return true;
+  return false;
 }
 
-export function SelectableMarkdownText(props: MobileSelectableMarkdownTextProps) {
-  return <T3SelectableMarkdownText {...props} highlightCode={highlightCodeSnippet} />;
+export function SelectableMarkdownText(_props: MobileSelectableMarkdownTextProps) {
+  return null;
 }
